add active admin files for precompressing

+++ lib/tasks/prepare_git_repo.rake
@@ -1,10 +1,16 @@
 namespace :deploy do
   task :before_deploy, [:env, :branch, :commitMessage] => :environment do |t, args|
-		puts "removing public assets..."
-		`git rm -rf public/assets/*`
-		`rm -rf public/assets/*`
-		puts "precompiling assets..."
-		`bundle exec rake assets:precompile`
+
+  	print "Do you want to precompile assets? (y/n) " and STDOUT.flush
+    char = $stdin.getc
+    if char != ?y && char != ?Y
+     	puts "removing public assets..."
+			`git rm -rf public/assets/*`
+			`rm -rf public/assets/*`
+			puts "precompiling assets..."
+			`bundle exec rake assets:precompile`
+    end
+
 		puts "add to git"
 		`git add .`
 		puts "commit to git with message: '#{args[:commitMessage]}'"
